fix(entcode,celt): prevent divide-by-zero and infinite normalize loops
cwrsV now uses uint64 arithmetic saturated at MaxUint32, so codebook
sizes never wrap back to zero for large (n,k) inputs.  Previously
icwrs(n,k) could return 0 via uint32 overflow, causing Decode(0) to
do rng/0 and panic.

- pvq.go: cwrsV returns uint64 with MaxUint32 saturation; encode/decode
  recursive splitting uses uint64 throughout; guards bail out when
  total >= cwrsMax instead of proceeding with a degenerate codebook.
- rate.go: computePulseCount uses uint64 maxIndex; terminates when
  cwrsV saturates (>= cwrsMax) so bits>=32 budgets no longer loop to
  the safety cap and produce oversized pulse counts.
- decoder.go: Decode() guards ft==0 and r==0; normalize() exits on
  rng==0 to prevent the infinite left-shift loop.
- encoder.go: Encode() guards r==0; normalize() exits on rng==0.

+++ internal/celt/celt_test.go
@@ -111,7 +111,7 @@ func TestBinomial(t *testing.T) {
 func TestCWRSV(t *testing.T) {
 	tests := []struct {
 		n, k int
-		want uint32
+		want uint64
 	}{
 		{1, 0, 1},
 		{1, 1, 2},
@@ -146,7 +146,7 @@ func TestCWRSIRoundtrip(t *testing.T) {
 	for _, c := range cases {
 		n, k := c[0], c[1]
 		total := cwrsV(n, k)
-		for idx := uint32(0); idx < total; idx++ {
+		for idx := uint32(0); uint64(idx) < total; idx++ {
 			y := cwrsi(n, k, idx)
 			// Check L1 norm
 			l1 := 0
